src/core/cicd_integration.py
import yaml
import logging

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class CICDIntegration:
    def __init__(self, config_path):
        with open(config_path, 'r') as file:
            self.config = yaml.safe_load(file)

    def check_policy_compliance(self, policy_id):
        # Sample data for policy compliance check
        sample_data = {
            'policy_id': policy_id,
            'status': 'compliant',
            'details': 'All checks passed successfully.'
        }
        logger.info(f"Policy {policy_id} compliance check completed: {sample_data['status']}")
        return sample_data

    def trigger_automated_check(self, policy_id):
        # Simulate triggering an automated check
        logger.info(f"Triggering automated check for policy {policy_id}")
        return self.check_policy_compliance(policy_id)

# Example usage
if __name__ == '__main__':
    cicd = CICDIntegration('config/settings.yaml')
    result = cicd.trigger_automated_check('availability-slo')
    print(result) 